refactor: split optics and shared UI scaffolding
