+++ src/components/layout/lensDiagram/AnalysisDrawerContent.tsx
@@ -1,10 +1,5 @@
 import { useDeferredValue, useEffect, useMemo, useRef, type ReactNode } from "react";
-import AberrationsPanel from "../../display/AberrationsPanel.js";
-import ComaTab from "../../display/ComaTab.js";
-import DistortionTab from "../../display/DistortionTab.js";
-import FocusBreathingTab from "../../display/FocusBreathingTab.js";
-import PupilAberrationTab from "../../display/PupilAberrationTab.js";
-import VignettingTab from "../../display/VignettingTab.js";
+import { ANALYSIS_TAB_RENDERERS } from "./analysisTabRenderers.js";
 import type { RuntimeLens } from "../../../types/optics.js";
 import type { Theme } from "../../../types/theme.js";
 import type { FieldGeometryState } from "../../../optics/optics.js";
@@ -92,89 +87,14 @@ export default function AnalysisDrawerContent({
     </>
   );
 
-  if (activeTab === "aberrations") {
-    return withMovementNotice(
-      <AberrationsPanel
-        L={L}
-        t={t}
-        zPos={zPos}
-        focusT={analysisInputs.focusT}
-        zoomT={analysisInputs.zoomT}
-        currentEPSD={analysisInputs.currentEPSD}
-        currentPhysStopSD={analysisInputs.currentPhysStopSD}
-        expanded={aberrationsExpanded}
-        onExpandedChange={onAberrationsExpandedChange}
-      />,
-    );
-  }
-
-  if (activeTab === "coma") {
-    return withMovementNotice(
-      <ComaTab
-        L={L}
-        t={t}
-        zPos={zPos}
-        focusT={analysisInputs.focusT}
-        zoomT={analysisInputs.zoomT}
-        currentEPSD={analysisInputs.currentEPSD}
-        currentPhysStopSD={analysisInputs.currentPhysStopSD}
-      />,
-    );
-  }
-
-  if (activeTab === "distortion") {
-    return withMovementNotice(
-      <DistortionTab
-        L={L}
-        t={t}
-        zPos={zPos}
-        focusT={analysisInputs.focusT}
-        zoomT={analysisInputs.zoomT}
-        dynamicEFL={analysisInputs.dynamicEFL}
-        currentPhysStopSD={analysisInputs.currentPhysStopSD}
-        fieldGeometry={analysisInputs.fieldGeometry}
-      />,
-    );
-  }
-
-  if (activeTab === "breathing") {
-    return withMovementNotice(
-      <FocusBreathingTab
-        L={L}
-        t={t}
-        focusT={analysisInputs.focusT}
-        zoomT={analysisInputs.zoomT}
-        dynamicEFL={analysisInputs.dynamicEFL}
-      />,
-    );
-  }
-
-  if (activeTab === "vignetting") {
-    return withMovementNotice(
-      <VignettingTab
-        L={L}
-        t={t}
-        zPos={zPos}
-        focusT={analysisInputs.focusT}
-        zoomT={analysisInputs.zoomT}
-        currentEPSD={analysisInputs.currentEPSD}
-        currentPhysStopSD={analysisInputs.currentPhysStopSD}
-        fieldGeometry={analysisInputs.fieldGeometry}
-      />,
-    );
-  }
-
-  if (activeTab === "pupils") {
-    return withMovementNotice(
-      <PupilAberrationTab
-        L={L}
-        t={t}
-        focusT={analysisInputs.focusT}
-        zoomT={analysisInputs.zoomT}
-        fieldGeometry={analysisInputs.fieldGeometry}
-      />,
-    );
-  }
-
-  return null;
+  return withMovementNotice(
+    ANALYSIS_TAB_RENDERERS[activeTab]({
+      L,
+      t,
+      zPos,
+      inputs: analysisInputs,
+      aberrationsExpanded,
+      onAberrationsExpandedChange,
+    }),
+  );
 }
