+++ src/components/layout/lensDiagram/analysisTabRenderers.tsx
@@ -1,0 +1,88 @@
+import type { ReactNode } from "react";
+import AberrationsPanel from "../../display/AberrationsPanel.js";
+import ComaTab from "../../display/ComaTab.js";
+import DistortionTab from "../../display/DistortionTab.js";
+import FocusBreathingTab from "../../display/FocusBreathingTab.js";
+import PupilAberrationTab from "../../display/PupilAberrationTab.js";
+import VignettingTab from "../../display/VignettingTab.js";
+import type { FieldGeometryState } from "../../../optics/optics.js";
+import type { RuntimeLens } from "../../../types/optics.js";
+import type { AnalysisTabId } from "../../../types/state.js";
+import type { Theme } from "../../../types/theme.js";
+
+export interface AnalysisDrawerInputs {
+  focusT: number;
+  zoomT: number;
+  currentEPSD: number;
+  currentPhysStopSD: number;
+  dynamicEFL: number;
+  fieldGeometry?: FieldGeometryState | null;
+}
+
+export interface AnalysisTabRendererContext {
+  L: RuntimeLens;
+  t: Theme;
+  zPos: number[];
+  inputs: AnalysisDrawerInputs;
+  aberrationsExpanded: boolean;
+  onAberrationsExpandedChange: (expanded: boolean) => void;
+}
+
+type AnalysisTabRenderer = (context: AnalysisTabRendererContext) => ReactNode;
+
+export const ANALYSIS_TAB_RENDERERS: Record<AnalysisTabId, AnalysisTabRenderer> = {
+  aberrations: ({ L, t, zPos, inputs, aberrationsExpanded, onAberrationsExpandedChange }) => (
+    <AberrationsPanel
+      L={L}
+      t={t}
+      zPos={zPos}
+      focusT={inputs.focusT}
+      zoomT={inputs.zoomT}
+      currentEPSD={inputs.currentEPSD}
+      currentPhysStopSD={inputs.currentPhysStopSD}
+      expanded={aberrationsExpanded}
+      onExpandedChange={onAberrationsExpandedChange}
+    />
+  ),
+  coma: ({ L, t, zPos, inputs }) => (
+    <ComaTab
+      L={L}
+      t={t}
+      zPos={zPos}
+      focusT={inputs.focusT}
+      zoomT={inputs.zoomT}
+      currentEPSD={inputs.currentEPSD}
+      currentPhysStopSD={inputs.currentPhysStopSD}
+    />
+  ),
+  distortion: ({ L, t, zPos, inputs }) => (
+    <DistortionTab
+      L={L}
+      t={t}
+      zPos={zPos}
+      focusT={inputs.focusT}
+      zoomT={inputs.zoomT}
+      dynamicEFL={inputs.dynamicEFL}
+      currentPhysStopSD={inputs.currentPhysStopSD}
+      fieldGeometry={inputs.fieldGeometry}
+    />
+  ),
+  breathing: ({ L, t, inputs }) => (
+    <FocusBreathingTab L={L} t={t} focusT={inputs.focusT} zoomT={inputs.zoomT} dynamicEFL={inputs.dynamicEFL} />
+  ),
+  vignetting: ({ L, t, zPos, inputs }) => (
+    <VignettingTab
+      L={L}
+      t={t}
+      zPos={zPos}
+      focusT={inputs.focusT}
+      zoomT={inputs.zoomT}
+      currentEPSD={inputs.currentEPSD}
+      currentPhysStopSD={inputs.currentPhysStopSD}
+      fieldGeometry={inputs.fieldGeometry}
+    />
+  ),
+  pupils: ({ L, t, inputs }) => (
+    <PupilAberrationTab L={L} t={t} focusT={inputs.focusT} zoomT={inputs.zoomT} fieldGeometry={inputs.fieldGeometry} />
+  ),
+};
