+++ src/optics/opticsFormat.ts
@@ -1,0 +1,20 @@
+import type { RuntimeLens } from "../types/optics.js";
+import { FOCUS_INFINITY_THRESHOLD } from "./layout.js";
+
+export function formatDist(t: number, L: RuntimeLens): string {
+  if (t < FOCUS_INFINITY_THRESHOLD) return "\u221e";
+  const d = L.closeFocusM / t;
+  if (d >= 100) return `${Math.round(d)} m`;
+  if (d >= 10) return `${d.toFixed(1)} m`;
+  if (d >= 1) return `${d.toFixed(2)} m`;
+  return `${(d * 100).toFixed(0)} cm`;
+}
+
+export function formatPetzvalRadius(P: number, subscript = true): string {
+  const label = subscript ? "R\u209a\u209c\u2093" : "R";
+  if (Math.abs(P) < 1e-6) return `${label} = \u221e`;
+  const R = 1 / P;
+  const absR = Math.abs(R);
+  const formatted = absR < 10 ? absR.toFixed(1) : Math.round(absR).toString();
+  return `${label} = ${R < 0 ? "\u2212" : ""}${formatted} mm`;
+}
